requirements.txt
pytest==8.3.3
pytest-playwright==0.5.2
pytest-html==4.1.1

